requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
autopep8==2.0.4
black==24.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.1
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-appconf==1.0.6
django-leaflet==0.29.0
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==23.1.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
phonenumberslite==8.13.30
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
rcssmin==1.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
yapf==0.40.2
zipp==3.17.0
